Fix PR comment
