+++ Harmony/Core/DependencyKeys.swift
@@ -5,8 +5,9 @@
 //  Created by Glny Gl on 08/02/2025.
 //
 
-import Dependencies
 import AVFoundation
+import Dependencies
+import SharingGRDB
 
 extension DependencyValues {
 
@@ -25,7 +26,7 @@ extension DependencyValues {
     set { self[MusicPlayerKey.self] = newValue }
   }
 
-  var favoriteService: FavoriteServiceProtocol {
+  var favoriteService: FavoriteService {
     get { self[FavoriteServiceKey.self] }
     set { self[FavoriteServiceKey.self] = newValue }
   }
@@ -44,7 +45,32 @@ private enum MusicPlayerKey: DependencyKey {
 }
 
 private enum FavoriteServiceKey: DependencyKey {
-  static var liveValue: FavoriteServiceProtocol {
-        return FavoriteService()
+  static var liveValue: FavoriteService {
+    @Dependency(\.defaultDatabase) var dbQueue
+
+    return FavoriteService(
+      addFavorite: { item in
+        let favorite = item.toFavoriteTrack()
+        try await dbQueue.write { db in
+          try favorite.insert(db)
+        }
+      },
+      deleteFavorite: { item in
+        try await dbQueue.write { db in
+          _ = try FavoriteTrack
+            .filter(Column("id") == item.id)
+            .deleteAll(db)
+        }
+      },
+      getFavorites: { limit, offset in
+        try await dbQueue.read { db in
+          let favorites = try FavoriteTrack
+            .order(Column("id").desc)
+            .limit(limit, offset: offset)
+            .fetchAll(db)
+          return favorites.map { $0.toTrackResponse() }
+        }
+      }
+    )
   }
 }
